feat: split tests unit unit and E2E.
The E2E tests running Closure are substantially
slower and mostly serve as only a sanity check.
