+++ test/sickle_test.ts
@@ -4,22 +4,12 @@ import * as ts from 'typescript';
 import {expect} from 'chai';
 
 import {annotateProgram, formatDiagnostics} from '../src/sickle';
-import {expectSource, checkClosureCompile} from './test_support';
+import {expectSource, goldenTests} from './test_support';
 
 describe('golden tests', () => {
-  var tsExtRe = /\.ts$/;
-  var testFolder = path.join(__dirname, '..', '..', 'test_files');
-  var files = fs.readdirSync(testFolder).filter((fn) => !!fn.match(tsExtRe));
-  var jsMasters: string[] = [];
-  files.forEach((tsFilename) => {
-    var tsPath = path.join(testFolder, tsFilename);
-    var tsSource = fs.readFileSync(tsPath, 'utf-8');
-    var jsPath = tsPath.replace(tsExtRe, '.js');
-    jsMasters.push(jsPath);
-    var jsSource = fs.readFileSync(jsPath, 'utf-8');
-    it(tsFilename, () => { expectSource(tsSource).to.equal(jsSource); });
+  goldenTests().forEach((test) => {
+    var tsSource = fs.readFileSync(test.tsPath, 'utf-8');
+    var jsSource = fs.readFileSync(test.jsPath, 'utf-8');
+    it(test.name, () => { expectSource(tsSource).to.equal(jsSource); });
   });
-
-  it('generates correct Closure code',
-     (done: (err: Error) => void) => { checkClosureCompile(jsMasters, done); });
 });
